[add] 処理の追加 - ロケーション機能 - 検索時と現在地の分岐 - その他
■Todo
- 通信の処理をリファクタリング（冗長すぎる？）
- UIの調整
- 日本語で検索できるようにする
-

+++ ClimaApp-JsonReEx/Model/WeatherData.swift
@@ -19,7 +19,7 @@ struct AcquisitionTargetAtWeatherItem: Codable {
 }
 
 struct AcquisitionTargetAtMainItem: Codable {
-    let temp: Double
+    let temp: Float
 }
 
 
@@ -70,7 +70,7 @@ struct AllWeatherItem: Codable {
 }
 
 struct AllMain: Codable {
-    let temp: Double
+    let temp: Float
     let feels_like: Double
     let temp_min: Double
     let temp_max: Double
